% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinatePointsMap_html.R
\name{coordinatePointsMap_html}
\alias{coordinatePointsMap_html}
\title{Plot coordinate pairs onto Google Maps HTML}
\usage{
coordinatePointsMap_html(latitude, longitude, data_frame, api_key = NULL,
  plot = TRUE, output_html_file = NULL)
}
\arguments{
\item{latitude}{List of latitudes. Required.}

\item{longitude}{List of longitudes. Required.}

\item{data_frame}{Entire data frame. Required.}

\item{plot}{Whether to open the HTML in a web browser. Defaults to TRUE.}

\item{output_html_file}{Write HTML map to this filepath if not null. Defaults to NULL.}
}
\description{
Plots given longitudes and latitudes onto a terrain map of the area. Pass a list as the "point_color" and / or "point_size" arguments in order to represent data graphically.
}
\examples{
tlm_data <- parsePayloadData("NS57_parsedPackets.txt", "LINK-TLM")
coordinatePointsMap_html(tlm_data$Latitude, tlm_data$Longitude, tlm_data)
}

