% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinatePointsMap_static.R
\name{coordinatePointsMap_static}
\alias{coordinatePointsMap_static}
\title{Plot coordinate pairs onto static map using ggplot2}
\usage{
coordinatePointsMap_static(latitudes, longitudes, point_color = "red",
  zoom = 10)
}
\arguments{
\item{latitudes}{List of latitudes. Required.}

\item{longitudes}{List of longitudes. Required.}

\item{point_color}{Color of points. Set to a constant color, or to a list to represent data graphically. Defaults to "red".}

\item{zoom}{Zoom level of map, range 1 to 21. Defaults to 10.}
}
\description{
Plots given longitudess and latitudess onto a terrain map of the area. Pass a list as the "point_color" and / or "point_size" arguments in order to represent data graphically.
}
\examples{
coordinatePointsMap_static(
    NS57_LINK_TLM$Latitude,
    NS57_LINK_TLM$Longitude,
    point_color = NS57_LINK_TLM$Altitude_m
)
}
