% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georeference_images.R
\name{georeference_images}
\alias{georeference_images}
\title{Calculate corner coordinates of aerial images}
\usage{
georeference_images(mag_x, mag_y, mag_z, pitch, roll, yaw)
}
\arguments{
\item{mag_x}{Magetometer X data}

\item{mag_y}{Magetometer X data}

\item{mag_z}{Magetometer X data}

\item{pitch}{IMU pitch data}

\item{roll}{IMU pitch data}

\item{yaw}{IMU pitch data}
}
\description{
Obtain corner coordinates from magetometer and IMU data
}
\examples{

}
\author{
Zach
}
