% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Join datasets and interpolate missing fields}
\usage{
interpolate(data_1, data_2, by)
}
\arguments{
\item{data_1}{First dataset.}

\item{data_2}{Second dataset.}

\item{by}{Key with which to join datasets.}
}
\description{
Merges datasets, specifically from IRENE and LINK-TLM
}
\examples{
interpolated_dataset <- interpolate(NS57_LINK_TLM, NS57_IRENE)

}
