% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.interpolate.R
\name{join.interpolate}
\alias{join.interpolate}
\title{Join datasets and interpolate missing fields}
\usage{
join.interpolate(data_1, data_2, by = c("DateTime", "Data_Source", "Flight"),
  exclude = NULL)
}
\arguments{
\item{data_1}{First dataset.}

\item{data_2}{Second dataset.}

\item{by}{One or more keys with which to join datasets, ordered by priority.}

\item{exclude}{List of column names to exclude from interpolation.}
}
\description{
Merges datasets, for instance IRENE and LINK-TLM, and interpolates numeric data to replace NA.
}
\examples{
NS57.RadCoord <- join.interpolate(NS57.Coord, NS57.Rad, exclude = c("Reading", "Unit"))
plot(NS57.RadCoord$Reading, NS57.RadCoord$Altitude_m)

}
