% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinData_interpolate.R
\name{joinData_interpolate}
\alias{joinData_interpolate}
\title{Join datasets and interpolate missing fields}
\usage{
joinData_interpolate(data_1, data_2, key, interpolate = FALSE)
}
\arguments{
\item{data_1}{First dataset.}

\item{data_2}{Second dataset.}

\item{key}{Key with which to join datasets.}

\item{interpolate}{Wether to interpolate data. Defaults to FALSE.}
}
\description{
Merges datasets, specifically from IRENE and LINK-TLM
}
\examples{
tlm_data <- parsePayloadData("NS57_parsedPackets.txt", "LINK-TLM")
irene_data <- parsePayloadData("NS57LaunchData.txt", "IRENE")
joined_data <- joinData_interpolate(tlm_data, irene_data, "Timestamp")
joined_data_interpolated <-
     joinData_interpolate(tlm_data,
                         irene_data,
                         "Timestamp",
                         interpolate = TRUE
     )
}

