% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.points.R
\name{map.points}
\alias{map.points}
\title{Plot coordinate pairs onto Google map tile}
\usage{
map.points(location_data, zoom = 10)
}
\arguments{
\item{location_data}{Dataset with latitude longitude pairs.}

\item{zoom}{Zoom level of map, range 1 to 21, defaults to 10.}
}
\description{
Plots given longitudes and latitudes onto a Google map tile of the area.
}
\examples{
map.points(NS57.Coord)
}
