% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NS57.Coord.R
\docType{data}
\name{NS57.Coord}
\alias{NS57.Coord}
\title{LINK-TLM Data}
\format{An object of class \code{data.frame} with 192 rows and 10 columns.}
\usage{
NS57.Coord
}
\description{
Flight NS57, 2016-07-30: Location data from received APRS packets.
}
\keyword{datasets}
