% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NS57.Rad.R
\docType{data}
\name{NS57.Rad}
\alias{NS57.Rad}
\title{IRENE Data}
\format{An object of class \code{data.frame} with 99 rows and 5 columns.}
\usage{
NS57.Rad
}
\description{
Flight NS57, 2016-07-30: Radiation data from the Geiger counter payload IRENE.
}
\keyword{datasets}
