% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NS57_IRENE.R
\docType{data}
\name{NS57_IRENE}
\alias{NS57_IRENE}
\title{IRENE Data}
\format{An object of class \code{data.frame} with 358 rows and 3 columns.}
\usage{
NS57_IRENE
}
\description{
Dataset containing radiation data from the Geiger counter payload IRENE from balloon flight NS57.
}
\keyword{datasets}

