% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NS57_LINK_TLM.R
\docType{data}
\name{NS57_LINK_TLM}
\alias{NS57_LINK_TLM}
\title{LINK-TLM Data}
\format{An object of class \code{data.frame} with 192 rows and 8 columns.}
\usage{
NS57_LINK_TLM
}
\description{
Dataset containing location data from the main command payload of balloon flight NS57.
}
\keyword{datasets}
