% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePayloadData.R
\name{parsePayloadData}
\alias{parsePayloadData}
\title{parsePayloadData}
\usage{
parsePayloadData(data_file, data_source, launch_timezone = Sys.timezone())
}
\arguments{
\item{data_file}{Data file in CSV format.}

\item{data_source}{Source of data, can be LINK-TLM or IRENE}

\item{launch_timezone}{Timezone of launch. Run OlsonNames for a list of timezones. Defaults to system timezone.}
}
\description{
Parses and cleans data from IRENE given input data file.
}
\examples{
library(balloonDataAnaylsis)
tlm_data <- parsePayloadData("NS57_parsedPackets.txt", "LINK-TLM")
irene_data <- parsePayloadData("NS57LaunchData.txt", "IRENE")
}

