% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultiSeries_differentUnits.R
\name{plotMultiSeries_differentUnits}
\alias{plotMultiSeries_differentUnits}
\title{Plot multiple series with different units on the same axis}
\usage{
plotMultiSeries_differentUnits(x, y1, y2, domain = NULL, x_name = "x",
  x_unit = NULL, y1_name = "y1", y1_unit = NULL, y1_color = "red",
  y1_draw_line = FALSE, y2_name = "y2", y2_unit = NULL,
  y2_color = "blue", y2_draw_line = FALSE, title = NULL,
  legend_pos = "topleft")
}
\arguments{
\item{x}{Shared x data (key).}

\item{y1}{First series.}

\item{y2}{Second series.}

\item{domain}{Defaults to domain of x axis.}

\item{x_name}{Defaults to "x".}

\item{x_unit}{Defaults to NULL.}

\item{y1_name}{Defaults to "y1".}

\item{y1_unit}{Defaults to NULL.}

\item{y1_color}{Defaults to "red".}

\item{y1_draw_line}{Defaults to FALSE.}

\item{y2_name}{Defaults to "y2".}

\item{y2_unit}{Defaults to NULL.}

\item{y2_color}{Defaults to "blue".}

\item{y2_draw_line}{Defaults to FALSE.}

\item{title}{Title of plot. Defaults to "y1_name (y1_unit) and y2_name (y2_unit) vs x_name (x_unit)".}

\item{legend_pos}{Position of legend, NULL for no legend. Defaults to "topleft".}
}
\description{
Plots two series with different units that share a common key.
Works best with interpolated data; balloonMergeDatasets(y1, y2, interpolate = TRUE)
Inspired by http://stackoverflow.com/questions/6142944/how-can-i-plot-with-2-different-y-axes
}
\examples{
joined_data <- joinData_interpolate(NS57_LINK_TLM, NS57_IRENE, "Timestamp", interpolate = TRUE)
plotMultiSeries_differentUnits(
    joined_data$Timestamp,
    joined_data$Reading,
    joined_data$Altitude_m,
    domain = c(min(NS57_LINK_TLM$Timestamp), max(NS57_LINK_TLM$Timestamp)),
    x_name = "Time",
    x_unit = "24hr",
    y1_name = "Geiger Counter",
    y1_unit = "counts per minute",
    y2_name = "Altitude",
    y2_unit = "meters",
    y2_draw_line = TRUE,
    legend_pos = NULL
)
}

