% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultiSeries_lengths.R
\name{plotMultiSeries_lengths}
\alias{plotMultiSeries_lengths}
\title{plotMultiSeries_lengths}
\usage{
plotMultiSeries_lengths(x, y1, y2, domain = NULL, range = NULL,
  x_axis_units = "x", y_axis_units = "y", y1_name = "y1",
  y1_color = "red", y2_name = "y2", y2_color = "blue", title = NULL)
}
\arguments{
\item{x}{Shared x data (key).}

\item{y1}{First series.}

\item{y2}{Second series.}

\item{domain}{Domain of plot.}

\item{range}{Range of plot.}

\item{x_axis_units}{Defaults to "x".}

\item{y_axis_units}{Defaults to "y".}

\item{y1_name}{Defaults to "y1".}

\item{y1_color}{Defaults to "red".}

\item{y2_name}{Defaults to "y2".}

\item{y2_color}{Defaults to "blue".}
}
\description{
Plots two series with the same unit (with possibly dissimilar dataset sizes) on the same graph using linear approximation
Works best with outer joined data; balloonMergeDatasets(y1, y2)
}
\examples{
library(balloonDataAnaylsis)
tlm_data <- balloonParseData("NS57_parsedPackets.txt", "LINK-TLM")
plotMultiSeries_lengths(
    tlm_data$Timestamp,
    tlm_data$Ascent_Rate_m_s,
    tlm_data$Ground_Speed_m_s,
    x_axis_units = "Time (24hr)",
    y_axis_units = "meters per second",
    y1_name = "Ascent Rate",
    y2_name = "Ground Speed"
)
}

