% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultiSeries_sameUnit.R
\name{plotMultiSeries_sameUnit}
\alias{plotMultiSeries_sameUnit}
\title{Plot multiple series with the same unit but possibly different dataset sizes on the same axis}
\usage{
plotMultiSeries_sameUnit(x, y1, y2, domain = NULL, range = NULL,
  x_axis_units = "x", y_axis_units = "y", y1_label = NULL,
  y1_color = "red", y2_label = NULL, y2_color = "blue", title = NULL,
  legend_pos = "topleft")
}
\arguments{
\item{x}{Shared x data (key).}

\item{y1}{First series.}

\item{y2}{Second series.}

\item{domain}{Domain of plot.}

\item{range}{Range of plot.}

\item{x_axis_units}{Defaults to "x".}

\item{y_axis_units}{Defaults to "y".}

\item{y1_label}{Defaults to "y1".}

\item{y1_color}{Defaults to "red".}

\item{y2_label}{Defaults to "y2".}

\item{y2_color}{Defaults to "blue".}

\item{title}{Defaults to "y1_label and y2_label vs x_axis_units"}

\item{legend_pos}{Position of legend, NULL for no legend. Defaults to "topleft".}
}
\description{
Plots two series with the same unit (with possibly dissimilar dataset sizes) on the same graph using linear approximation
Works best with outer joined data; balloonMergeDatasets(y1, y2)
}
\examples{
plotMultiSeries_sameUnit(
    NS57_LINK_TLM$DateTime,
    NS57_LINK_TLM$Ascent_Rate_m_s,
    NS57_LINK_TLM$Ground_Speed_m_s,
    x_axis_units = "Time (24hr)",
    y_axis_units = "meters per second",
    y1_label = "Ascent Rate",
    y2_label = "Ground Speed",
    legend_pos = "bottomleft"
)
}
