% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.payload.R
\name{read.payload}
\alias{read.payload}
\title{Read UMDBPP payload data from file}
\usage{
read.payload(file, data_source, flight_number = NULL, flight_date = NULL,
  start_time = NULL, end_time = NULL, timezone = Sys.timezone())
}
\arguments{
\item{file}{Path to log file.}

\item{data_source}{Source from which data originated, from the list c("LINK-TLM", "CellTracker", "APRS", "IRENE").}

\item{flight_number}{Flight designation.}

\item{flight_date}{Date of flight in "YYYY-MM-DD" format. Defaults to NULL.}

\item{start_time}{Time in 24 hour "HH:MM:SS" format. If a time is given, dataset will be cut to after start time.}

\item{end_time}{Time in 24 hour "HH:MM:SS" format. If a time is given, dataset will be cut to before end time.}

\item{timezone}{Timezone of launch. Run OlsonNames for a list of accepted timezones.}
}
\value{
Returns data frame of parsed payload data, standardized to POSIXct timestamps and SI units.
}
\description{
This function will read and parse data University of Maryland Balloon Payload Program (UMDBPP) data logs.
It standardizes units and calculates ascent rate, ground speed, and downrange distance per entry.
}
\examples{
\dontrun{
NS57.Coord <- read.payload("NS57_parsedPackets.txt",
        data_source = "LINK-TLM",
        flight_number = "NS57",
        start_time = "07:50:31",
        end_time = "09:27:34"
)

NS57.Rad <- read.payload("NS57LaunchData.txt",
        data_source = "IRENE",
        flight_number = "NS57",
        start_time = "07:50:31",
        end_time = "09:27:34"
)
}
}
