% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinatePointsMap_html.R
\name{coordinatePointsMap_html}
\alias{coordinatePointsMap_html}
\title{Plot coordinate pairs onto Google Maps HTML}
\usage{
coordinatePointsMap_html(latitudes, longitudes, data_frame, api_key = NULL,
  plot = TRUE, output_html_file = NULL)
}
\arguments{
\item{data_frame}{Data frame that must at least contain the fields "Latitude" and "Longitude". Required.}

\item{api_key}{Google Maps API Key. Required if you want to host the HTML output on a server. Defaults to NULL.}

\item{plot}{Whether to open map in browser. Defaults to FALSE.}

\item{output_html_file}{Name of output file. Defaults to NULL for no file writing.}
}
\description{
Plots given longitudes and latitudes onto a terrain map of the area. Pass a list as the "point_color" and / or "point_size" arguments in order to represent data graphically.
}
\examples{
coordinatePointsMap_html(NS57_LINK_TLM)
}

