% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinatePointsMap_static.R
\name{coordinatePointsMap_static}
\alias{coordinatePointsMap_static}
\title{Plot coordinate pairs onto static map using ggplot2}
\usage{
coordinatePointsMap_static(latitudes, longitudes, point_color = "red",
  point_size = 3, zoom_level = 10)
}
\arguments{
\item{latitudes}{List of latitudes. Required.}

\item{longitudes}{List of longitudes. Required.}

\item{point_color}{Color of points. Set to a constant color, or to a list to represent data graphically. Defaults to "red".}

\item{point_size}{Size of points. Set to a constant size, or to a list to represent data graphically. Defaults to 2.}

\item{zoom_level}{Zoom level of map. 3 is world, 10 is city, 21 is street. May have to fine tune this variable to get good map. Defaults to 10.}

\item{title}{Map title. Set to element_blank() for none. Defaults to "Altitude (meters)".}
}
\description{
Plots given longitudess and latitudess onto a terrain map of the area. Pass a list as the "point_color" and / or "point_size" arguments in order to represent data graphically.
}
\examples{
coordinatePointsMap_static(
    NS57_LINK_TLM$Latitude,
    NS57_LINK_TLM$Longitude,
    point_color = NS57_LINK_TLM$Altitude_m,
    title = "Altitude (meters)"
)
}

