% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.interpolate.R
\name{join.interpolate}
\alias{join.interpolate}
\title{Join datasets and interpolate missing fields}
\usage{
join.interpolate(data_1, data_2, by = c("DateTime", "Data_Source", "Flight"))
}
\arguments{
\item{data_1}{First dataset.}

\item{data_2}{Second dataset.}

\item{by}{One or more key with which to join datasets, ordered by priority.}
}
\description{
Merges datasets, for instance IRENE and LINK-TLM
}
\examples{
interpolated_dataset <- join.interpolate(NS57.Coord, NS57.Rad)

}
