% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinData_interpolate.R
\name{joinData_interpolate}
\alias{joinData_interpolate}
\title{Join datasets and interpolate missing fields}
\usage{
joinData_interpolate(data_1, data_2, by, interpolate = TRUE)
}
\arguments{
\item{data_1}{First dataset.}

\item{data_2}{Second dataset.}

\item{by}{Key with which to join datasets.}

\item{interpolate}{Wether to interpolate data. Defaults to TRUE.}
}
\description{
Merges datasets, specifically from IRENE and LINK-TLM
}
\examples{
joined_data <- joinData_interpolate(NS57_LINK_TLM, NS57_IRENE, interpolate = FALSE)
joined_data_interpolated <- joinData_interpolate(NS57_LINK_TLM, NS57_IRENE)

}

