% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePayloadData.R
\name{parsePayloadData}
\alias{parsePayloadData}
\title{Parse UMDBPP payload data from file}
\usage{
parsePayloadData(data_file, payload = c("LINK-TLM", "IRENE", "CellTracker"),
  launch_timezone = Sys.timezone())
}
\arguments{
\item{data_file}{Data file in CSV format.}

\item{payload}{Payload from which data originated. Defaults to "LINK-TLM".}

\item{launch_timezone}{Timezone of launch. Run OlsonNames for a list of timezones. Defaults to system timezone.}
}
\value{
Returns data frame of parsed payload data, standardized to POSIXct timestamps and SI units.
}
\description{
Parses and cleans data from given data file, assuming format of given payload.
}
\examples{
\dontrun{
tlm_data <- parsePayloadData("NS57_parsedPackets.txt", "LINK-TLM")
irene_data <- parsePayloadData("NS57LaunchData.txt", "IRENE")
}
}

