% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultiSeries_differentUnits.R
\name{plotMultiSeries_differentUnits}
\alias{plotMultiSeries_differentUnits}
\title{Plot multiple series with different units on the same axis}
\usage{
plotMultiSeries_differentUnits(x, y, xlim = NULL, xlab = NULL,
  ylab = NULL, type = "p", main = NULL)
}
\arguments{
\item{x}{X series.}

\item{y}{Y series.}

\item{xlim}{Defaults to domain of x axis.}

\item{xlab}{X labels.}

\item{ylab}{Y labels.}

\item{type}{Defaults to FALSE.}

\item{main}{Title of plot. Defaults to NULL.}
}
\description{
Plots two series with different units that share a common key.
Works best with interpolated data; balloonMergeDatasets(y_left, y2, interpolate = TRUE)
Inspired by http://stackoverflow.com/questions/6142944/how-can-i-plot-with-2-different-y-axes
}
\examples{
plotMultiSeries_differentUnits(
    x = list(NS57_IRENE$DateTime, NS57_LINK_TLM$DateTime),
    y = list(NS57_IRENE$Reading, NS57_LINK_TLM$Altitude_m),
    xlim = c(min(NS57_LINK_TLM$DateTime), max(NS57_LINK_TLM$DateTime)),
    xlab = c("Time", "Time"),
    ylab = c("Geiger Counter (CPM)", "Altitude (m)"),
    type = "l"
)
}
